move region-based construction into `TableData` class
